+++ kitty/boss.py
@@ -484,6 +484,7 @@ class Boss:
             sz = os_window_font_size(os_window_id)
             if sz:
                 os_window_font_size(os_window_id, sz, True)
+                tm.update_dpi_based_sizes()
                 tm.resize()
 
     def _set_os_window_background_opacity(self, os_window_id, opacity):